+++ LabSoftware VS2010/Colour.h
@@ -7,7 +7,7 @@ public:
 	Colour();
 	Colour(unsigned char r, unsigned char g, unsigned char b);
 	Colour(const Colour &other);
-	Colour Interpolate(Colour colour, int steps, int index) const;
+	Colour Interpolate(Colour colour, float steps, float index) const;
 
 	bool operator==(const Colour & otherColour) const;
 };
